# swagger.yaml
openapi: 3.0.0
info:
  title: Психологическая Служба API
  description: API для сервиса психологической помощи
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Локальный сервер разработки
  - url: http://185.128.105.126:8000
    description: Сервер разработки
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Описание ошибки
...
    AppointmentBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        therapist_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [Online, Offline]
        reason:
          type: string
          nullable: true
        status:
          type: string
          enum: [Accepted, Cancelled, Done]
        remind_time:
          type: string
          format: date-time
        last_change_time:
          type: string
          format: date-time
        venue:
          type: string
    AppointmentCreateRequest:
      type: object
      required:
        - patient_id
        - therapist_id
        - type
      properties:
        patient_id:
          type: string
          format: uuid
        therapist_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [Online, Offline]
        reason:
          type: string
          nullable: true
        remind_time:
          type: string
          format: date-time
          nullable: true
        venue:
          type: string
          nullable: true
    AppointmentCreateResponse:
      type: object
      properties:
        appointment_id:
          type: string
          format: uuid
...
paths:
...
  /appointments/{appointment_id}:
    get:
      tags:
        - appointments
      summary: Получить информацию о записи по ID
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentBase'
        '404':
          description: Запись не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appointments/create:
    post:
      tags:
        - appointments
      summary: Создать новую запись
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '200':
          description: Запись успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentCreateResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /appointments/{appointment_id}/cancel:
    put:
      tags:
        - appointments
      summary: Отмена записи
      description: Отменяет существующую запись к психологу
      parameters:
        - name: appointment_id
          in: path
          required: true
          description: ID записи для отмены
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Запись успешно отменена
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Встреча успешно отменена"
        '404':
          description: Запись не найдена
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Встреча не найдена"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Внутренняя ошибка сервера"
...